Moved magento sync services to magento connector bundle.
